elasticsearch_movie_search/requirements.txt
elasticsearch==8.11.1
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy>=1.24.0
torch==2.8.0
